Address review comments from * #1885 * #1886
